modal window use render props
